app/requirements.txt
betterlogging==0.1.0
envreader==0.0.5
uvicorn==0.18.3
starlette==0.20.4
sh==1.14.3
pyjwt==2.4.0
click==8.1.3
requests==2.28.1
